refact: changes on identity classes
- change classes names on Identity
